Create Find_Value.lsp
This is the code used to find the property value and/or property tag in any block. A list of all of the matching results are both printed and returned to the user. This code is in a working state for me, but I also see potential for further improvement. The evaluation of every property in every block on a blank template drawing roughly takes two to three minutes.