Create Find_Value_Debugger.lsp
In the case that the default function does not work as intended, this file is designed to allow the implementer a way to track and solve the issue themselves. This file expects the user to have some awareness on coding and debugging. The results are printed to a text file located under "C:\Users\[User Name]\Appdata\Local\Temp"  with the name of "AutoLisp-Print.txt."